// caspargus_be/src/models/CaspargusEvent.ts
export interface ICaspargusEventSummary {
  _id: string;
  createdAt: Date;
  updatedAt: Date;
  name: string;
}




/* export class CaspargusEvent {
  _id!: string;
  createdAt: Date = new Date();
  updatedAt: Date = new Date();
  name: string;

  constructor(name: string) {
    this.name = name;
  }
} */